Change long datatype for labels to uint8

+++ unsat/data.py
@@ -82,7 +82,7 @@ class XRayDataset(Dataset):
         labels = self.hdf5_file[sample_name]['labels'][day_idx, height_idx]
 
         data = torch.from_numpy(data).type(torch.float32)
-        labels = torch.from_numpy(labels).type(torch.long)
+        labels = torch.from_numpy(labels).type(torch.uint8)
 
         return data, labels
